backend/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
